refactor: error modal, notification bar
