+++ src/components/index.js
@@ -9,6 +9,8 @@ export { default as AddressInput } from "./inputs/addressInput";
 export { default as CheckBox } from './Checkbox'
 export { default as Steppers } from './Steppers'
 export { default as ConfirmPopup } from './ConfirmPopup'
+export { default as ErrorPopup } from './ErrorPopup'
+export { default as NotificationBar } from './NotificationBar'
 
 export { default as DisplaySeed } from "./displaySeed/displaySeed";
 export { default as ConfirmSeed } from "./confirmSeed/confirmSeed";
